Recipe owner is not allowed to save/unsave recipe

+++ server/src/services/savedRecipesService.ts
@@ -9,6 +9,30 @@ export function savedRecipesService(database: Database) {
   const recipesRepository = buildRecipesRepository(database);
 
   return {
+    async create(userId: string, recipeId: number) {
+      const recipeById = await recipesRepository.findById(recipeId);
+
+      if (!recipeById)
+        throw new TRPCError({
+          code: 'NOT_FOUND',
+          message: 'Recipe was not found!',
+        });
+
+      if (recipeById.userId === userId) {
+        throw new TRPCError({
+          code: 'BAD_REQUEST',
+          message: 'You cannot save your own recipe!',
+        });
+      }
+
+      const createdRecipe = await savedRecipesRepository.create({
+        userId,
+        recipeId,
+      });
+
+      return createdRecipe;
+    },
+
     async remove(recipeId: number, userId: string) {
       const recipeById = await recipesRepository.findById(recipeId);
 
@@ -17,6 +41,13 @@ export function savedRecipesService(database: Database) {
           code: 'NOT_FOUND',
           message: 'Recipe was not found!',
         });
+
+      if (recipeById.userId === userId) {
+        throw new TRPCError({
+          code: 'BAD_REQUEST',
+          message: 'You cannot unsave your own recipe!',
+        });
+      }
 
       try {
         const unsavedRecipe = await savedRecipesRepository.remove(
